PR(ADHERE): Adhere to code review - move to utils
