+++ http/middleware.go
@@ -26,63 +26,6 @@ import (
 	"github.com/sourcenetwork/defradb/internal/db"
 )
 
-const (
-	// txHeaderName is the name of the transaction header.
-	// This header should contain a valid transaction id.
-	txHeaderName = "x-defradb-tx"
-)
-
-type contextKey string
-
-var (
-	// txsContextKey is the context key for the transaction *sync.Map
-	txsContextKey = contextKey("txs")
-	// dbContextKey is the context key for the client.DB
-	dbContextKey = contextKey("db")
-	// colContextKey is the context key for the client.Collection
-	//
-	// If a transaction exists, all operations will be executed
-	// in the current transaction context.
-	colContextKey = contextKey("col")
-)
-
-// mustGetContextClientCollection returns the client collection from the http request context or panics.
-//
-// This should only be called from functions within the http package.
-func mustGetContextClientCollection(req *http.Request) client.Collection {
-	return req.Context().Value(colContextKey).(client.Collection) //nolint:forcetypeassert
-}
-
-// mustGetContextSyncMap returns the sync map from the http request context or panics.
-//
-// This should only be called from functions within the http package.
-func mustGetContextSyncMap(req *http.Request) *sync.Map {
-	return req.Context().Value(txsContextKey).(*sync.Map) //nolint:forcetypeassert
-}
-
-// mustGetContextClientDB returns the client DB from the http request context or panics.
-//
-// This should only be called from functions within the http package.
-func mustGetContextClientDB(req *http.Request) client.DB {
-	return req.Context().Value(dbContextKey).(client.DB) //nolint:forcetypeassert
-}
-
-// mustGetContextClientStore returns the client store from the http request context or panics.
-//
-// This should only be called from functions within the http package.
-func mustGetContextClientStore(req *http.Request) client.Store {
-	return req.Context().Value(dbContextKey).(client.Store) //nolint:forcetypeassert
-}
-
-// tryGetContextClientP2P returns the P2P client from the http request context and a boolean
-// indicating if p2p was enabled.
-//
-// This should only be called from functions within the http package.
-func tryGetContextClientP2P(req *http.Request) (client.P2P, bool) {
-	p2p, ok := req.Context().Value(dbContextKey).(client.P2P)
-	return p2p, ok
-}
-
 // CorsMiddleware handles cross origin request
 func CorsMiddleware(allowedOrigins []string) func(http.Handler) http.Handler {
 	return cors.Handler(cors.Options{
